Move kernels into separate source file and create them only when needed
